rest_scraper/requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.3
django-extensions==2.2.8
djangorestframework==3.11.0
django-filter==2.2.0
idna==2.9
Pillow==7.0.0
psycopg2-binary==2.8.4
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
redis
celery==4.3.0
# https://github.com/celery/kombu/issues/1063
kombu==4.6.3
selectolax==0.2.3
